Move agency general information into separate file
